macos: make learn-image-shas didn't work on mac
On macOS, due to the vendor-go target, when running

  make learn-tools-shas

the command would fail with:

  bash: line 1: /var/tmp/.../_bin/tools/go: cannot execute binary file: Exec format error

That's because it runs the command:

  ./scripts/learn_tools_shas.sh tools vendor-go

which in turn was calling:

  make \
   HOST_OS=linux \
   HOST_ARCH=arm64 \
    -f ./learn_tools_shas.helper.mk tools vendor-go

Since vendor-go was used, the wrong Go binary was used. On macOS, you can only
use the darwin/arm64 binary, which led to the Exec format error.

A solution to work around this is to run "make vendor-go" beforehand, and then
pretend that vendor-go had been given by using GO=_bin/tools/go.

Signed-off-by: Maël Valais <mael@vls.dev>